requirements.txt
# -*- conf-mode -*-
coverage>=3.7.1
#cssselect>=0.6.1               # for PyQuery
decorator>=3.4.0
defusedxml>=0.4.1		# for TastyPie when ussing xml; not a declared dependency
django-bootstrap3>=5.1.1
django-widget-tweaks>=1.3
docutils>=0.12                  
html5lib>=0.90
#lxml>=3.4.0                    # for PyQuery; 
mimeparse>=0.1.3		# for TastPie
#MySQL-python>=1.2.5
pathlib>=1.0
pip>=6.0
pyflakes>=0.8.1
pyquery>1.2.4
python-dateutil>=2.2
python-magic>=0.4.6
pytz>=2014.7
setuptools>=1.2
six>=1.8.0
django-tastypie>=0.12.1
wsgiref>=0.1.2
